refactor CategoryCreationPage
